+++ src/main/java/gift/product/infrastructure/persistence/repository/JpaProductRepository.java
@@ -1,5 +1,6 @@
-package gift.product.infrastructure.persistence;
+package gift.product.infrastructure.persistence.repository;
 
+import gift.product.infrastructure.persistence.entity.ProductEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaProductRepository extends JpaRepository<ProductEntity, Long> {
