+++ src/main/java/gift/wishes/infrastructure/persistence/WishesRepositoryImpl.java
@@ -6,13 +6,12 @@ import gift.core.domain.product.exception.ProductNotFoundException;
 import gift.core.domain.user.User;
 import gift.core.domain.user.exception.UserNotFoundException;
 import gift.core.domain.wishes.WishesRepository;
-import gift.product.infrastructure.persistence.JpaProductRepository;
-import gift.product.infrastructure.persistence.ProductEntity;
+import gift.product.infrastructure.persistence.repository.JpaProductRepository;
+import gift.product.infrastructure.persistence.entity.ProductEntity;
 import gift.user.infrastructure.persistence.JpaUserRepository;
 import gift.user.infrastructure.persistence.UserEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 import java.util.List;
