+++ src/test/java/gift/product/ProductRepositoryTests.java
@@ -1,15 +1,13 @@
 package gift.product;
 
-import gift.product.infrastructure.persistence.JpaProductCategoryRepository;
-import gift.product.infrastructure.persistence.JpaProductRepository;
-import gift.product.infrastructure.persistence.ProductCategoryEntity;
-import gift.product.infrastructure.persistence.ProductEntity;
-import org.junit.jupiter.api.BeforeEach;
+import gift.product.infrastructure.persistence.repository.JpaProductCategoryRepository;
+import gift.product.infrastructure.persistence.repository.JpaProductRepository;
+import gift.product.infrastructure.persistence.entity.ProductCategoryEntity;
+import gift.product.infrastructure.persistence.entity.ProductEntity;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.test.context.jdbc.Sql;
 
 import static org.assertj.core.api.Assertions.assertThat;
