+++ src/test/java/gift/wishes/WishRepositoryTests.java
@@ -1,9 +1,9 @@
 package gift.wishes;
 
-import gift.product.infrastructure.persistence.JpaProductCategoryRepository;
-import gift.product.infrastructure.persistence.JpaProductRepository;
-import gift.product.infrastructure.persistence.ProductCategoryEntity;
-import gift.product.infrastructure.persistence.ProductEntity;
+import gift.product.infrastructure.persistence.repository.JpaProductCategoryRepository;
+import gift.product.infrastructure.persistence.repository.JpaProductRepository;
+import gift.product.infrastructure.persistence.entity.ProductCategoryEntity;
+import gift.product.infrastructure.persistence.entity.ProductEntity;
 import gift.user.infrastructure.persistence.JpaUserRepository;
 import gift.user.infrastructure.persistence.UserEntity;
 import gift.wishes.infrastructure.persistence.JpaWishRepository;
@@ -13,7 +13,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.test.context.jdbc.Sql;
 
 import static org.assertj.core.api.Assertions.assertThat;
